pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dnsintel"
version = "0.1.0"
description = "A comprehensive DNS intelligence and reconnaissance tool"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["dns", "whois", "ssl", "security", "reconnaissance"]
requires-python = ">=3.8"
dependencies = [
    "dnspython>=2.3.0",
    "python-whois>=0.8.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "colorama>=0.4.6",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
dnsintel = "dnsintel.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/dnsintel"
Documentation = "https://github.com/yourusername/dnsintel#readme"
Repository = "https://github.com/yourusername/dnsintel"
"Bug Tracker" = "https://github.com/yourusername/dnsintel/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dnsintel*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

